Fix using backspace to delete elements
Closes #54
Closes #53

+++ src/js/prototypes/atom.js
@@ -191,9 +191,10 @@ export default class Atom {
     }
     
     keyPress(key){
-        //runs whenver a key is pressed
-        if (key == 'Delete'){
-            if(this.selected == true){
+        //runs whenever a key is pressed
+        if (['Delete', 'Backspace'].includes(key)){
+            if(this.selected == true && document.getElementsByTagName("BODY")[0] == document.activeElement){
+                //If this atom is selected AND the body is active (meaning we are not typing in a text box)
                 this.deleteNode();
             }
         }
